INTR-73: Add support for Enterprise Model Scanner (#11)
* set up CI/CD pipeline and environment for testing against enterprise model scanner

* add support for enterprise model scanner

* cleanup code

* add env vars needed in upcoming model scanner release. switch to use latest

+++ core/src/modelScannerPlugin.groovy
@@ -86,7 +86,7 @@ download {
                 }
                 log.debug "file: $responseRepoPath status: $modelStatus"
                 repositories.setProperty(responseRepoPath, 'hiddenlayer.status', modelStatus)
-                if (config.deleteAfterScan) {
+                if (config.deleteAfterScan && api.isSaaS()) {
                     sensorId = modelScanner.getSensorIdForUrl(modelInfo.repoPath)
                     api.deleteModel(sensorId)
                 }
